Added the category and migrated using redux
