+++ src/redux/constants/categoriesConstant.js
@@ -1,0 +1,3 @@
+export const CATEGORIES_ACTION_TYPES = {
+  SET_CATEGORIES_MAP: 'SET_CATEGORIES_MAP',
+};
