+++ src/redux/reducers/index.js
@@ -1,6 +1,8 @@
 import { combineReducers } from 'redux';
 import { userReducer } from './user/userReducer';
+import { categoriesReducer } from './categories/categoriesReducer';
 
 export const rootReducer = combineReducers({
   user: userReducer,
+  categories: categoriesReducer,
 });
