--- src/services/UserContext.jsx
@@ -1,52 +1,0 @@
-import React, { createContext, useEffect, useReducer } from 'react';
-import {
-  onAuthStateChangedListener,
-  createUserDocumentFromAuth,
-} from '../utils/firebase/Firebase.config';
-
-export const UserContext = createContext({
-  currentUser: null,
-  setCurrentUser: () => null,
-});
-
-export const USER_ACTION_TYPES = {
-  SET_CURRENT_USER: 'SET_CURRENT_USER',
-};
-
-const userReducer = (state, action) => {
-  const { type, payload } = action;
-
-  switch (type) {
-    case USER_ACTION_TYPES.SET_CURRENT_USER:
-      return {
-        ...state,
-        currentUser: payload,
-      };
-    default:
-      throw new Error(`unHandled type ${type} in userReducer`);
-  }
-};
-
-const initialState = {
-  currentUser: null,
-};
-
-export const UserProvider = ({ children }) => {
-  const [{ currentUser }, dispatch] = useReducer(userReducer, initialState);
-  const setCurrentUser = (user) => {
-    dispatch({ type: USER_ACTION_TYPES.SET_CURRENT_USER, payload: user });
-  };
-
-  const value = { currentUser, setCurrentUser };
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChangedListener((user) => {
-      if (user) {
-        createUserDocumentFromAuth(user);
-      }
-      setCurrentUser(user);
-    });
-    return unsubscribe;
-  }, []);
-  return <UserContext.Provider value={value}>{children} </UserContext.Provider>;
-};
